BIG set of changes! This changes the implementation of the "filter" method by moving most of the work out of the functions and into the %in% method, which is exposed to the user. This is designed to be easier for programmers and future filter implementors.
filter() itself is now a single function that dispatches the appropriate %in%. Also, we're now carrying around some extra filterResult data to help us make filterResult a type of filter class that can be used in conjunction with other unresolved filters or with sets of filter results using the usual sorts of logical operators. 

We also have as(...,"logical") for filters and resolved filters (compositions of filterResults) can be converted to a single logical vector for more advanced subset manipulation.

I also added the beginning of a vignette that will outline using filters.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/flowCore@21406 bc3139a8-67e5-0310-9ffc-ced21a209358