As far as I can tell the argument "inv" to compensate is never used or tested in flowCore or plateCore.
Since it is never used, but if you change the value it will give incorrect answers, and since it has been a source of confusion, I thought it best to eliminate it.

I removed it from all function calls, and I commented out any functional lines of code that contained it
and replaced if necessary.

git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/flowCore@35819 bc3139a8-67e5-0310-9ffc-ced21a209358